Job filtering: by description, by location, by hours per week, by cost

+++ unijobs-core/src/main/java/core/service/FetchService.java
@@ -16,7 +16,7 @@ public interface FetchService {
     UniUser getOneById(Integer id);
 
     List<Job> getAllJobs();
-    Job findJob(Long id);
+    Job findJob(int id);
 
     List<UniUser> getAllUsers();
     List<TemporaryUser> getAllTemporaryUsers();
